update directory di controller dan views
