+++ app/Http/Controllers/AdmindanSuperadmin/Adminpage/ResepController.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace App\Http\Controllers\Adminpage;
+namespace App\Http\Controllers\AdmindanSuperadmin\Adminpage;
 
 
 use App\Http\Controllers\Controller;
-
-use App\Models\Article;
+use App\Models\Resep;
+use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
-class ArticleController extends Controller
+class ResepController extends Controller
 {
     public function index(Request $request)
     {
-        $query = Article::latest();
+        $query = Resep::latest();
 
         /* Melakukan Filter Data */
         if ($request->get('search')) {
@@ -25,17 +24,48 @@ class ArticleController extends Controller
 
         $data = $query->paginate(5);
     
-        return view('admindansuperadmin.adminpage.article.index', compact('data','request'));
+        return view('admindansuperadmin.adminpage.resep.index', compact('data','request'));
     }
 
+        // ResepController.php
+
+    public function show(Request $request, $id)
+    {
+            $data = Resep::find($id);
+
+            return view('admindansuperadmin.adminpage.resep.show', compact('data'));
+    }
+
+    /* Add Comment Untuk Resep */
+    public function addComment(Request $request, $id)
+    {
+        $request->validate([
+            'comment' => 'required',
+        ]);
+
+        $commentData = [
+            'user_id' => Auth::id(),
+            'resep_id' => $id,
+            'comment_resep' => $request->input('comment'),
+        ];
+
+        Comment::create($commentData);
+
+        return redirect()->back()->with('success', 'Komentar berhasil ditambahkan');
+    }
+
+
     public function create(){
-        return view('admindansuperadmin.adminpage.article.create');
+        return view('admindansuperadmin.adminpage.resep.create');
     }
 
     public function store(Request $request){
         $validator = Validator::make($request->all(),[
             'image' => 'required|mimes:png,jpg,jpeg|max:2048',
             'judul' => 'required', 
+            'waktu' => 'required',
+            'porsi' => 'required',
+            'bahan_langkah' => 'required',
             'description' => 'required',
         ]);
 
@@ -45,31 +75,40 @@ class ArticleController extends Controller
         /* Eksekusi Image */
         $image = $request->file('image');
         $filename = date('Y-m-d').$image->getClientOriginalName();
-        $path = 'photo-article/'.$filename;
+        $path = 'photo-resep/'.$filename;
 
         Storage::disk('public')->put($path,file_get_contents($image));
 
         /* Ngirim Data */
         $data['user_id'] = Auth::id();
         $data['judul']      = $request->judul;
-        $data['description']       = $request->description;
+        $data['waktu']      = $request->waktu;
+        $data['porsi']      = $request->porsi;
+        
+        // Decode HTML entities before saving to the database
+        $data['description'] = htmlspecialchars_decode($request->description);
+        $data['bahan_langkah'] = htmlspecialchars_decode($request->bahan_langkah);
+
         $data['image']      = $filename;
 
-        Article::create($data);
+        Resep::create($data);
 
-        return redirect()->route('admin.article.index');
+        return redirect()->route('admin.resep.index');
     }
 
     public function edit(Request $request, $id){
-        $data = Article::find($id); 
+        $data = Resep::find($id); 
         
-        return view('admindansuperadmin.adminpage.article.edit', compact('data'));
+        return view('admindansuperadmin.adminpage.resep.edit', compact('data'));
     }
 
     public function update(Request $request, $id){
         $validator = Validator::make($request->all(),[
             'image' => 'nullable|mimes:png,jpg,jpeg|max:2048',
             'judul' => 'required', 
+            'waktu' => 'required',
+            'porsi' => 'required',
+            'bahan_langkah' => 'required',
             'description' => 'required',
         ]);
     
@@ -79,14 +118,17 @@ class ArticleController extends Controller
         }
     
         /* Ngirim Data */
-        $data = Article::find($id);
+        $data = Resep::find($id);
         $data->judul = $request->judul;
+        $data->waktu = $request->waktu;
+        $data->porsi = $request->porsi;
+        $data->bahan_langkah = $request->bahan_langkah;
         $data->description = $request->description;
 
          // Periksa apakah kotak centang "remove_image" dipilih
         if ($request->has('remove_image')) {
             // Hapus gambar lama jika ada
-            Storage::disk('public')->delete('photo-article/' . $data->image);
+            Storage::disk('public')->delete('photo-resep/' . $data->image);
             $data->image = null;
         }
     
@@ -95,7 +137,7 @@ class ArticleController extends Controller
             // Simpan gambar yang baru
             $image = $request->file('image');
             $filename = date('Y-m-d') . $image->getClientOriginalName();
-            $path = 'photo-article/' . $filename;
+            $path = 'photo-resep/' . $filename;
             Storage::disk('public')->put($path, file_get_contents($image));
             $data->image = $filename;
         }
@@ -103,17 +145,18 @@ class ArticleController extends Controller
 
         $data->save(); // Simpan pengguna yang telah diperbarui
     
-        return redirect()->route('admin.article.index');
+        return redirect()->route('admin.resep.index');
     }
 
     public function delete(Request $request, $id){
-        $data = Article::find($id);
+        $data = Resep::find($id);
 
         if($data){
             $data->delete();
         }
 
-        return redirect()->route('admin.article.index');
+        return redirect()->route('admin.resep.index');
     }
     
+
 }
