+++ app/Http/Controllers/AdmindanSuperadmin/Superadminpage/DataTableController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Superadminpage;
+namespace App\Http\Controllers\AdmindanSuperadmin\Superadminpage;
 
 
 use App\Http\Controllers\Controller;
