+++ resources/views/admindansuperadmin/superadminpage/user/edit.blade.php
@@ -1,4 +1,4 @@
-@extends('admindansuperadminsuperadminpage.layouts.main')
+@extends('admindansuperadmin.superadminpage.layouts.main')
 @section('content')
 
 <div class="content-wrapper">
